Rename HaskellDynamicClusteringTest to HaskellRuntimeTest
